plus multi, single mode, rankingsystem, separate header,function
