+++ MultiPlay.h
@@ -1,0 +1,8 @@
+#ifndef __MULTPLAY_H__
+#define __MULTPLAY_H__
+#include "Function.h"
+#define BOARD_SIZE 5 // 게임 보드의 크기 정의
+#define SHIPS_COUNT 4 // 게임의 전함 수 정의
+
+void multplay();
+#endif
