+++ SinglePlay.h
@@ -1,0 +1,9 @@
+#ifndef __SINGLEPLAY_H__
+#define __SINGLEPLAY_H__
+#include "Function.h"
+#include "Ranking.h"
+#define BOARD_SIZE 5 // 게임 보드의 크기 정의
+#define SHIPS_COUNT 4 // 게임의 전함 수 정의
+void singleplay();
+
+#endif
